updated SiaUtil , partialy working SiaHostedDownloadTest, better error messages

+++ minebd/src/test/java/io/minebox/nbd/SiaHostedDownloadTest.java
@@ -1,18 +1,23 @@
 package io.minebox.nbd;
 
+import ch.qos.logback.classic.Level;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import io.minebox.SiaUtil;
 import net.lingala.zip4j.core.ZipFile;
 import org.apache.commons.io.FileUtils;
 import org.junit.*;
+import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 
+@Ignore //we need a solution for the big pristine file
 public class SiaHostedDownloadTest {
 
     private static Process siadProcess;
+    private static File siadDirectoryFile;
+    private static SiaUtil siaUtil;
 
     private static <T> T doTimed(Callable<T> runnable) {
         final long start = System.currentTimeMillis();
@@ -25,6 +30,11 @@ public class SiaHostedDownloadTest {
             System.out.println("finished with error in " + (System.currentTimeMillis() - start) + " ms");
             throw new RuntimeException(e);
         }
+    }
+
+    static {
+        final ch.qos.logback.classic.Logger logger = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger("org.apache.http");
+        logger.setLevel(Level.INFO);
     }
 
     private static File recreatePristineSiad() throws IOException {
@@ -46,33 +56,48 @@ public class SiaHostedDownloadTest {
 
     @BeforeClass
     public static void setupSiad() throws IOException {
+        siaUtil = new SiaUtil("http://localhost:9980");
+//        siadDirectoryFile = recreatePristineSiad();
+        siadDirectoryFile = new File("junit/sia/Sia-v1.3.0-linux-amd64/");
+//        startSiad(siadDirectoryFile);
+    }
 
-        final File siadDirectoryFile = recreatePristineSiad();
-
+    private static void startSiad(File siadDirectoryFile) throws IOException {
         siadProcess = new ProcessBuilder("./siad")
                 .directory(siadDirectoryFile)
                 .inheritIO()
                 .start();
+        siaUtil.waitForConsensus();
+        siaUtil.unlockWallet("rage return onto madness abort vegan under inwardly madness stick swept tucks demonstrate duration viking oneself extra jabbed arsenic dotted renting orchid honked mighty onslaught batch tugs jagged absorb");
     }
 
     @AfterClass
     public static void tearDownSiad() throws Exception {
-        siadProcess.destroyForcibly();
+        stopSiad();
+    }
+
+    private static void stopSiad() {
+        siaUtil.gracefulStop();
+        try {
+            Thread.sleep(1000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        if (siadProcess != null) {
+            siadProcess.destroyForcibly();
+        }
     }
 
 
     @Test
     public void testInit() throws Exception {
-        final SiaUtil siaUtil = new SiaUtil("http://localhost:9980");
-        siaUtil.waitForConsensus();
-        siaUtil.unlockWallet("rage return onto madness abort vegan under inwardly madness stick swept tucks demonstrate duration viking oneself extra jabbed arsenic dotted renting orchid honked mighty onslaught batch tugs jagged absorb");
-
         final RemoteTokenService dummyRemoteToken = new RemoteTokenService(null, null) {
             @Override
             public Optional<String> getToken() {
                 return Optional.of("123");
             }
         };
+        //acutally, we want to make sure, sia is not running here...
         final SiaHostedDownload underTest = new SiaHostedDownload(siaUtil,
                 null,
                 "junit/sia/Sia-v1.3.0-linux-amd64",
@@ -81,16 +106,33 @@ public class SiaHostedDownloadTest {
             @Override
             protected InputStream downloadLatestMetadataZip(String token) throws UnirestException {
                 try {
-                    return new FileInputStream("/home/andreas/minebox/backup.1503060902.zip");
+                    return new FileInputStream("/home/andreas/minebox/backup.1503407403.zip");
                 } catch (FileNotFoundException e) {
                     throw new RuntimeException(e);
                 }
             }
 
+            @Override
+            protected void finishedDigest() {
+                restartSiad();
+            }
         };
         underTest.initKeyListener();
-        final boolean result = underTest.downloadIfPossible(new File("minebox_v1_0.dat"));
+        for (String filename : underTest.allFilenames()) {
+            final File dest = new File("/tmp/sia/", filename);
+            System.out.println("attempting restore of: " + filename + " to destination " + dest.getAbsolutePath());
+            final DownloadService.RecoveryStatus recoveryStatus = underTest.downloadIfPossible(dest);
+            System.out.println(dest.getAbsolutePath() + " recovery: " + recoveryStatus);
+        }
+    }
 
-
+    private void restartSiad() {
+        stopSiad();
+        try {
+            startSiad(siadDirectoryFile);
+        } catch (IOException e) {
+            Assert.fail(e.getMessage());
+            throw new RuntimeException(e);
+        }
     }
 }
